Move MessageEventDispatcher into plugin as well
